third_party/blink/web_tests/external/wpt/css/css-shapes/shape-outside/values/shape-outside-ellipse-002-expected.txt

This is a testharness.js-based test.
Found 136 tests; 106 PASS, 30 FAIL, 0 TIMEOUT, 0 NOTRUN.
[PASS] ellipse(at 50%) serializes as ellipse(at 50% 50%) - inline
[PASS] ellipse(at 50px) serializes as ellipse(at 50px 50%) - inline
[PASS] ellipse(at 50% 50%) serializes as ellipse(at 50% 50%) - inline
[PASS] ellipse(at 50% 50px) serializes as ellipse(at 50% 50px) - inline
[PASS] ellipse(at 50px 50%) serializes as ellipse(at 50px 50%) - inline
[PASS] ellipse(at 50px 50px) serializes as ellipse(at 50px 50px) - inline
[PASS] ellipse(at left) serializes as ellipse(at 0% 50%) - inline
[PASS] ellipse(at top) serializes as ellipse(at 50% 0%) - inline
[PASS] ellipse(at right) serializes as ellipse(at 100% 50%) - inline
[PASS] ellipse(at bottom) serializes as ellipse(at 50% 100%) - inline
[PASS] ellipse(at center) serializes as ellipse(at 50% 50%) - inline
[PASS] ellipse(at left top) serializes as ellipse(at 0% 0%) - inline
[PASS] ellipse(at left bottom) serializes as ellipse(at 0% 100%) - inline
[PASS] ellipse(at left center) serializes as ellipse(at 0% 50%) - inline
[PASS] ellipse(at top left) serializes as ellipse(at 0% 0%) - inline
[PASS] ellipse(at top right) serializes as ellipse(at 100% 0%) - inline
[PASS] ellipse(at top center) serializes as ellipse(at 50% 0%) - inline
[PASS] ellipse(at right top) serializes as ellipse(at 100% 0%) - inline
[PASS] ellipse(at right bottom) serializes as ellipse(at 100% 100%) - inline
[PASS] ellipse(at right center) serializes as ellipse(at 100% 50%) - inline
[PASS] ellipse(at bottom left) serializes as ellipse(at 0% 100%) - inline
[PASS] ellipse(at bottom right) serializes as ellipse(at 100% 100%) - inline
[PASS] ellipse(at bottom center) serializes as ellipse(at 50% 100%) - inline
[PASS] ellipse(at center top) serializes as ellipse(at 50% 0%) - inline
[PASS] ellipse(at center left) serializes as ellipse(at 0% 50%) - inline
[PASS] ellipse(at center right) serializes as ellipse(at 100% 50%) - inline
[PASS] ellipse(at center bottom) serializes as ellipse(at 50% 100%) - inline
[PASS] ellipse(at center center) serializes as ellipse(at 50% 50%) - inline
[PASS] ellipse(at left 50%) serializes as ellipse(at 0% 50%) - inline
[PASS] ellipse(at left 50px) serializes as ellipse(at 0% 50px) - inline
[PASS] ellipse(at 50% top) serializes as ellipse(at 50% 0%) - inline
[PASS] ellipse(at 50px top) serializes as ellipse(at 50px 0%) - inline
[PASS] ellipse(at right 80%) serializes as ellipse(at 100% 80%) - inline
[PASS] ellipse(at right 80px) serializes as ellipse(at 100% 80px) - inline
[PASS] ellipse(at 70% bottom) serializes as ellipse(at 70% 100%) - inline
[PASS] ellipse(at 70px bottom) serializes as ellipse(at 70px 100%) - inline
[PASS] ellipse(at center 60%) serializes as ellipse(at 50% 60%) - inline
[PASS] ellipse(at center 60px) serializes as ellipse(at 50% 60px) - inline
[PASS] ellipse(at 60% center) serializes as ellipse(at 60% 50%) - inline
[PASS] ellipse(at 60px center) serializes as ellipse(at 60px 50%) - inline
[PASS] ellipse(at left 50% top 50%) serializes as ellipse(at 50% 50%) - inline
[PASS] ellipse(at left 50% top 50px) serializes as ellipse(at 50% 50px) - inline
[FAIL] ellipse(at left 50% bottom 70%) serializes as ellipse(at 50% calc(30%)) - inline
  assert_equals: expected "ellipse(at 50% calc(30%))" but got "ellipse(at 50% 30%)"
[FAIL] ellipse(at left 50% bottom 70px) serializes as ellipse(at 50% calc(100% - 70px)) - inline
  assert_equals: expected "ellipse(at 50% calc(100% - 70px))" but got "ellipse(at left 50% bottom 70px)"
[PASS] ellipse(at left 50px top 50%) serializes as ellipse(at 50px 50%) - inline
[PASS] ellipse(at left 50px top 50px) serializes as ellipse(at 50px 50px) - inline
[FAIL] ellipse(at left 50px bottom 70%) serializes as ellipse(at 50px calc(30%)) - inline
  assert_equals: expected "ellipse(at 50px calc(30%))" but got "ellipse(at 50px 30%)"
[PASS] ellipse(at top 50% left 50%) serializes as ellipse(at 50% 50%) - inline
[PASS] ellipse(at top 50% left 50px) serializes as ellipse(at 50px 50%) - inline
[FAIL] ellipse(at top 50% right 80%) serializes as ellipse(at calc(20%) 50%) - inline
  assert_equals: expected "ellipse(at calc(20%) 50%)" but got "ellipse(at 20% 50%)"
[FAIL] ellipse(at top 50% right 80px) serializes as ellipse(at calc(100% - 80px) 50%) - inline
  assert_equals: expected "ellipse(at calc(100% - 80px) 50%)" but got "ellipse(at right 80px top 50%)"
[PASS] ellipse(at top 50px left 50%) serializes as ellipse(at 50% 50px) - inline
[PASS] ellipse(at top 50px left 50px) serializes as ellipse(at 50px 50px) - inline
[FAIL] ellipse(at top 50px right 80%) serializes as ellipse(at calc(20%) 50px) - inline
  assert_equals: expected "ellipse(at calc(20%) 50px)" but got "ellipse(at 20% 50px)"
[FAIL] ellipse(at bottom 70% left 50%) serializes as ellipse(at 50% calc(30%)) - inline
  assert_equals: expected "ellipse(at 50% calc(30%))" but got "ellipse(at 50% 30%)"
[FAIL] ellipse(at bottom 70% left 50px) serializes as ellipse(at 50px calc(30%)) - inline
  assert_equals: expected "ellipse(at 50px calc(30%))" but got "ellipse(at 50px 30%)"
[FAIL] ellipse(at bottom 70% right 80%) serializes as ellipse(at calc(20%) calc(30%)) - inline
  assert_equals: expected "ellipse(at calc(20%) calc(30%))" but got "ellipse(at 20% 30%)"
[FAIL] ellipse(at bottom 70% right 80px) serializes as ellipse(at calc(100% - 80px) calc(30%)) - inline
  assert_equals: expected "ellipse(at calc(100% - 80px) calc(30%))" but got "ellipse(at right 80px top 30%)"
[FAIL] ellipse(at bottom 70px left 50%) serializes as ellipse(at 50% calc(100% - 70px)) - inline
  assert_equals: expected "ellipse(at 50% calc(100% - 70px))" but got "ellipse(at left 50% bottom 70px)"
[FAIL] ellipse(at bottom 70px right 50%) serializes as ellipse(at calc(50%) calc(100% - 70px)) - inline
  assert_equals: expected "ellipse(at calc(50%) calc(100% - 70px))" but got "ellipse(at left 50% bottom 70px)"
[FAIL] ellipse(at bottom 70px right 80px) serializes as ellipse(at calc(100% - 80px) calc(100% - 70px)) - inline
  assert_equals: expected "ellipse(at calc(100% - 80px) calc(100% - 70px))" but got "ellipse(at right 80px bottom 70px)"
[FAIL] ellipse(at right 80% top 50%) serializes as ellipse(at calc(20%) 50%) - inline
  assert_equals: expected "ellipse(at calc(20%) 50%)" but got "ellipse(at 20% 50%)"
[FAIL] ellipse(at right 80% top 50px) serializes as ellipse(at calc(20%) 50px) - inline
  assert_equals: expected "ellipse(at calc(20%) 50px)" but got "ellipse(at 20% 50px)"
[FAIL] ellipse(at right 80% bottom 70%) serializes as ellipse(at calc(20%) calc(30%)) - inline
  assert_equals: expected "ellipse(at calc(20%) calc(30%))" but got "ellipse(at 20% 30%)"
[FAIL] ellipse(at right 80% bottom 70px) serializes as ellipse(at calc(20%) calc(100% - 70px)) - inline
  assert_equals: expected "ellipse(at calc(20%) calc(100% - 70px))" but got "ellipse(at left 20% bottom 70px)"
[FAIL] ellipse(at right 80px top 50%) serializes as ellipse(at calc(100% - 80px) 50%) - inline
  assert_equals: expected "ellipse(at calc(100% - 80px) 50%)" but got "ellipse(at right 80px top 50%)"
[FAIL] ellipse(at right 80px bottom 70%) serializes as ellipse(at calc(100% - 80px) calc(30%)) - inline
  assert_equals: expected "ellipse(at calc(100% - 80px) calc(30%))" but got "ellipse(at right 80px top 30%)"
[FAIL] ellipse(at right 80px bottom 70px) serializes as ellipse(at calc(100% - 80px) calc(100% - 70px)) - inline
  assert_equals: expected "ellipse(at calc(100% - 80px) calc(100% - 70px))" but got "ellipse(at right 80px bottom 70px)"
[PASS] ellipse(at 50%) serializes as ellipse(at 50% 50%) - computed
[PASS] ellipse(at 50px) serializes as ellipse(at 50px 50%) - computed
[PASS] ellipse(at 50% 50%) serializes as ellipse(at 50% 50%) - computed
[PASS] ellipse(at 50% 50px) serializes as ellipse(at 50% 50px) - computed
[PASS] ellipse(at 50px 50%) serializes as ellipse(at 50px 50%) - computed
[PASS] ellipse(at 50px 50px) serializes as ellipse(at 50px 50px) - computed
[PASS] ellipse(at left) serializes as ellipse(at 0% 50%) - computed
[PASS] ellipse(at top) serializes as ellipse(at 50% 0%) - computed
[PASS] ellipse(at right) serializes as ellipse(at 100% 50%) - computed
[PASS] ellipse(at bottom) serializes as ellipse(at 50% 100%) - computed
[PASS] ellipse(at center) serializes as ellipse(at 50% 50%) - computed
[PASS] ellipse(at left top) serializes as ellipse(at 0% 0%) - computed
[PASS] ellipse(at left bottom) serializes as ellipse(at 0% 100%) - computed
[PASS] ellipse(at left center) serializes as ellipse(at 0% 50%) - computed
[PASS] ellipse(at top left) serializes as ellipse(at 0% 0%) - computed
[PASS] ellipse(at top right) serializes as ellipse(at 100% 0%) - computed
[PASS] ellipse(at top center) serializes as ellipse(at 50% 0%) - computed
[PASS] ellipse(at right top) serializes as ellipse(at 100% 0%) - computed
[PASS] ellipse(at right bottom) serializes as ellipse(at 100% 100%) - computed
[PASS] ellipse(at right center) serializes as ellipse(at 100% 50%) - computed
[PASS] ellipse(at bottom left) serializes as ellipse(at 0% 100%) - computed
[PASS] ellipse(at bottom right) serializes as ellipse(at 100% 100%) - computed
[PASS] ellipse(at bottom center) serializes as ellipse(at 50% 100%) - computed
[PASS] ellipse(at center top) serializes as ellipse(at 50% 0%) - computed
[PASS] ellipse(at center left) serializes as ellipse(at 0% 50%) - computed
[PASS] ellipse(at center right) serializes as ellipse(at 100% 50%) - computed
[PASS] ellipse(at center bottom) serializes as ellipse(at 50% 100%) - computed
[PASS] ellipse(at center center) serializes as ellipse(at 50% 50%) - computed
[PASS] ellipse(at left 50%) serializes as ellipse(at 0% 50%) - computed
[PASS] ellipse(at left 50px) serializes as ellipse(at 0% 50px) - computed
[PASS] ellipse(at 50% top) serializes as ellipse(at 50% 0%) - computed
[PASS] ellipse(at 50px top) serializes as ellipse(at 50px 0%) - computed
[PASS] ellipse(at right 80%) serializes as ellipse(at 100% 80%) - computed
[PASS] ellipse(at right 80px) serializes as ellipse(at 100% 80px) - computed
[PASS] ellipse(at 70% bottom) serializes as ellipse(at 70% 100%) - computed
[PASS] ellipse(at 70px bottom) serializes as ellipse(at 70px 100%) - computed
[PASS] ellipse(at center 60%) serializes as ellipse(at 50% 60%) - computed
[PASS] ellipse(at center 60px) serializes as ellipse(at 50% 60px) - computed
[PASS] ellipse(at 60% center) serializes as ellipse(at 60% 50%) - computed
[PASS] ellipse(at 60px center) serializes as ellipse(at 60px 50%) - computed
[PASS] ellipse(at left 50% top 50%) serializes as ellipse(at 50% 50%) - computed
[PASS] ellipse(at left 50% top 50px) serializes as ellipse(at 50% 50px) - computed
[PASS] ellipse(at left 50% bottom 70%) serializes as ellipse(at 50% 30%) - computed
[FAIL] ellipse(at left 50% bottom 70px) serializes as ellipse(at 50% calc(100% - 70px)) - computed
  assert_equals: expected "ellipse(at 50% calc(100% - 70px))" but got "ellipse(at left 50% bottom 70px)"
[PASS] ellipse(at left 50px top 50%) serializes as ellipse(at 50px 50%) - computed
[PASS] ellipse(at left 50px top 50px) serializes as ellipse(at 50px 50px) - computed
[PASS] ellipse(at left 50px bottom 70%) serializes as ellipse(at 50px 30%) - computed
[PASS] ellipse(at top 50% left 50%) serializes as ellipse(at 50% 50%) - computed
[PASS] ellipse(at top 50% left 50px) serializes as ellipse(at 50px 50%) - computed
[PASS] ellipse(at top 50% right 80%) serializes as ellipse(at 20% 50%) - computed
[FAIL] ellipse(at top 50% right 80px) serializes as ellipse(at calc(100% - 80px) 50%) - computed
  assert_equals: expected "ellipse(at calc(100% - 80px) 50%)" but got "ellipse(at right 80px top 50%)"
[PASS] ellipse(at top 50px left 50%) serializes as ellipse(at 50% 50px) - computed
[PASS] ellipse(at top 50px left 50px) serializes as ellipse(at 50px 50px) - computed
[PASS] ellipse(at top 50px right 80%) serializes as ellipse(at 20% 50px) - computed
[PASS] ellipse(at bottom 70% left 50%) serializes as ellipse(at 50% 30%) - computed
[PASS] ellipse(at bottom 70% left 50px) serializes as ellipse(at 50px 30%) - computed
[PASS] ellipse(at bottom 70% right 80%) serializes as ellipse(at 20% 30%) - computed
[FAIL] ellipse(at bottom 70% right 80px) serializes as ellipse(at calc(100% - 80px) 30%) - computed
  assert_equals: expected "ellipse(at calc(100% - 80px) 30%)" but got "ellipse(at right 80px top 30%)"
[FAIL] ellipse(at bottom 70px left 50%) serializes as ellipse(at 50% calc(100% - 70px)) - computed
  assert_equals: expected "ellipse(at 50% calc(100% - 70px))" but got "ellipse(at left 50% bottom 70px)"
[FAIL] ellipse(at bottom 70px right 50%) serializes as ellipse(at 50% calc(100% - 70px)) - computed
  assert_equals: expected "ellipse(at 50% calc(100% - 70px))" but got "ellipse(at left 50% bottom 70px)"
[FAIL] ellipse(at bottom 70px right 80px) serializes as ellipse(at calc(100% - 80px) calc(100% - 70px)) - computed
  assert_equals: expected "ellipse(at calc(100% - 80px) calc(100% - 70px))" but got "ellipse(at right 80px bottom 70px)"
[PASS] ellipse(at right 80% top 50%) serializes as ellipse(at 20% 50%) - computed
[PASS] ellipse(at right 80% top 50px) serializes as ellipse(at 20% 50px) - computed
[PASS] ellipse(at right 80% bottom 70%) serializes as ellipse(at 20% 30%) - computed
[FAIL] ellipse(at right 80% bottom 70px) serializes as ellipse(at 20% calc(100% - 70px)) - computed
  assert_equals: expected "ellipse(at 20% calc(100% - 70px))" but got "ellipse(at left 20% bottom 70px)"
[FAIL] ellipse(at right 80px top 50%) serializes as ellipse(at calc(100% - 80px) 50%) - computed
  assert_equals: expected "ellipse(at calc(100% - 80px) 50%)" but got "ellipse(at right 80px top 50%)"
[FAIL] ellipse(at right 80px bottom 70%) serializes as ellipse(at calc(100% - 80px) 30%) - computed
  assert_equals: expected "ellipse(at calc(100% - 80px) 30%)" but got "ellipse(at right 80px top 30%)"
[FAIL] ellipse(at right 80px bottom 70px) serializes as ellipse(at calc(100% - 80px) calc(100% - 70px)) - computed
  assert_equals: expected "ellipse(at calc(100% - 80px) calc(100% - 70px))" but got "ellipse(at right 80px bottom 70px)"
Harness: the test ran to completion.
